volk_common: math cores to own file; robustify C++ usage
As Greg Troxel figured out, isnan() might not be globally available on
every C++ compiler (where it might be std::isnan); same for isinf.

Anyways, a bunch of partially not even used math functions have no
business cluttering the volk_common.h header that contains basic
declarations and is included from everywhere. It has even less business
including math.h at that point!

Move that functionality, used by very few kernels, into volk_mathfun.h,
hopefully fix the compiler-doesn't-automatically-using std::isnan.

While doing that, also make sure `#define`d constants are UPPERCASE, and
all functions are `volk_` prefixed. This is C – we can't be cluttering
the one namespace we have.

Adding the header to the kernels that are using these functions, I
noticed that these headers have global includes not include-guarded.
That's probably a pretty bad idea, so fixed these kernels on the go.
