+++ kernels/volk/volk_32f_sin_32f.h
@@ -56,6 +56,7 @@
  *   volk_free(out);
  * \endcode
  */
+#include <volk/volk_mathfun.h>
 
 #include <inttypes.h>
 #include <math.h>
