+++ kernels/volk/volk_32fc_s32f_atan2_32f.h
@@ -61,6 +61,8 @@
 #ifndef INCLUDED_volk_32fc_s32f_atan2_32f_a_H
 #define INCLUDED_volk_32fc_s32f_atan2_32f_a_H
 
+#include <volk/volk_mathfun.h>
+
 #include <math.h>
 
 #ifdef LV_HAVE_GENERIC
